Fixed removing child element causing error
When removing a child element, an error was thrown saying `clientHeight` was `undefined` on `item.element`. The solution is to filter out items with a `null` element.

This issue is caused by `this.items` having the previous children after `componentDidUpdate` and not the updated children from `this.props.children`. This is a stop-gap fix until a real fix can be made which re-computes the children off `this.props.children` in `this.updateLayout`.